// backend/src/core/domain/enums/impl/error-codes.enums.ts
export enum ErrorCodes {
  CANNOT_ACTIVATE_ACCOUNT = 'CANNOT_ACTIVATE_ACCOUNT',
  ACCOUNT_NOT_FOUND = 'ACCOUNT_NOT_FOUND',

  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  ACCOUNT_NOT_CREATED = 'ACCOUNT_NOT_CREATED',
  USER_OR_PASSWORD_INVALID = 'USER_OR_PASSWORD_INVALID',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_NOT_CREATED = 'USER_NOT_CREATED',

  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  UNAUTHORIZED = 'UNAUTHORIZED',
  HASH_EXPIRED = 'HASH_EXPIRED',

  PLACE_NOT_FOUND = 'PLACE_NOT_FOUND',
  PLACE_NOT_CREATED = 'PLACE_NOT_CREATED',
  PLACE_MISSING_LAT_LNG = 'PLACE_MISSING_LAT_LNG',

  ITEM_NOT_FOUND = 'ITEM_NOT_FOUND',

  PLAYER_NOT_FOUND = 'PLAYER_NOT_FOUND',

  INVENTORY_NOT_FOUND = 'INVENTORY_NOT_FOUND',

  SETS_NOT_FOUND = 'SETS_NOT_FOUND',
  SET_NOT_CREATED = 'SET_NOT_CREATED',

  ITEM_NOT_CREATED = 'ITEM_NOT_CREATED',
  ITEM_NOT_UPDATED = 'ITEM_NOT_UPDATED',

  PARTY_NOT_FOUND = 'PARTY_NOT_FOUND',
  PLAYER_ALREADY_IN_PARTY = 'PLAYER_ALREADY_IN_PARTY',
  PARTY_MEMBERS_NOT_FOUND = 'PARTY_MEMBERS_NOT_FOUND',

  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',

  DUNGEON_NOT_FOUND = 'DUNGEON_NOT_FOUND',
}